fix(clipboard): Only use tmux provider if Nvim runs in side tmux
This essentially reverts 2495e7e (and documentation updates done later).
That past change meant that we would modify the buffer contents of a
tmux session if it exists, even if the current Nvim process wasn't
running inside of it. Depending on the tmux configuration, this could
even affect the clipboard of an actually attached tmux client, since
tmux itself uses OSC 52 to forward buffer writes to attached clients.

While autodetection is usually a trade-off and can rarely make everybody
happy, this behavior goes counter the principle of least surprise. If
really desired, it can be brought back by explicit configuration.